mcp_server/requirements.txt
pdf2image>=1.16.3
pytesseract>=0.3.10
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0